1.19: * rename message package
