+++ src/main/java/tfar/quickstack/networking/S2CReportPacket.java
@@ -1,4 +1,4 @@
-package tfar.quickstack.message;
+package tfar.quickstack.networking;
 
 import java.util.ArrayList;
 import java.util.List;
